requirements.txt
langchain
openai
streamlit
tiktoken
python-dotenv
docarray
langchain-community --upgrade